Add support to past own ID into forum threads
